universalized jijiij
